docs(Table): separate renderer & server utils
