+++ packages/core/src/Table/stories/TableRenderers/TagColumnRenderer.tsx
@@ -14,7 +14,7 @@ import {
   useHvPagination,
 } from "@hitachivantara/uikit-react-core";
 
-import { EmptyRow, makeRenderersData, NewRendererEntry } from "../storiesUtils";
+import { EmptyRow, makeRenderersData, NewRendererEntry } from "./utils";
 
 export const TagColumnRenderer = () => {
   const columns = useMemo(() => {
@@ -33,9 +33,7 @@ export const TagColumnRenderer = () => {
     ];
   }, []);
 
-  const initialData = useMemo(() => makeRenderersData(64), []);
-
-  const [data] = useState(initialData);
+  const [data] = useState(() => makeRenderersData(64));
 
   const {
     getTableProps,
