+++ packages/core/src/Table/stories/TableSamples/ListRow.tsx
@@ -1,4 +1,4 @@
-import { useMemo } from "react";
+import { useMemo, useState } from "react";
 import {
   HvButton,
   HvCheckBox,
@@ -13,24 +13,16 @@ import {
   HvTableRow,
 } from "@hitachivantara/uikit-react-core";
 
-import {
-  AssetEvent,
-  getColumns,
-  makeData,
-  useToggleIndex,
-} from "../storiesUtils";
+import { AssetEvent, getColumns, makeData } from "../storiesUtils";
 
 export const ListRow = () => {
-  const [checkedIdx, toggleChecked] = useToggleIndex(0);
+  const [checkedIdx, setCheckedIdx] = useState(0);
 
   const columns = useMemo<HvTableColumnConfig<AssetEvent, string>[]>(() => {
-    return [...getColumns(), { Header: "Details", accessor: "link" }];
+    return [...getColumns(), { Header: "Details", id: "link" }];
   }, []);
 
-  const data = useMemo<AssetEvent[]>(
-    () => makeData(4).map((row) => ({ ...row, link: "/details" })),
-    [],
-  );
+  const data = useMemo<AssetEvent[]>(() => makeData(4), []);
 
   if (!data) return null;
 
@@ -52,7 +44,7 @@ export const ListRow = () => {
                 <HvCheckBox
                   aria-label="Tick to select the row"
                   checked={checkedIdx === idx}
-                  onClick={toggleChecked(idx)}
+                  onClick={() => setCheckedIdx(idx === checkedIdx ? -1 : idx)}
                 />
               </HvTableCell>
               <HvTableCell>{el.name}</HvTableCell>
